+++ examples/next/pages/ui/components/storage/storage-browser/default-auth/routed/[locations]/[location-detail]/index.page.tsx
@@ -6,7 +6,6 @@ import { Button, Flex } from '@aws-amplify/ui-react';
 
 import { StorageBrowser } from '../../StorageBrowser';
 
-import '@aws-amplify/ui-react-storage/storage-browser-styles.css';
 import '@aws-amplify/ui-react-storage/styles.css';
 
 export default function Page() {