+++ packages/react-storage/src/components/StorageBrowser/actions/configs/index.ts
@@ -1,11 +1,13 @@
 export {
   ActionConfigsProvider,
   ActionConfigsProviderProps,
-  useActionConfig,
+  useActionConfigs,
 } from './context';
 export {
   defaultActionConfigs,
   defaultActionViewConfigs,
+  DefaultActionViewType,
   isDefaultActionViewType,
 } from './defaults';
 export * from './types';
+export { getActionConfigs } from './utils';