+++ packages/react-storage/src/components/StorageBrowser/index.ts
@@ -1,9 +1,21 @@
 export { componentsDefault } from './componentsDefault';
 export { createStorageBrowser } from './createStorageBrowser';
 export {
+  ActionViewConfig,
+  ActionHandler,
+  FileDataItem,
+  ExtendedActionConfigs,
+} from './actions';
+export {
   createAmplifyAuthAdapter,
   createManagedAuthAdapter,
   CreateManagedAuthAdapterInput,
   StorageBrowserAuthAdapter,
 } from './adapters';
-export { CreateStorageBrowserInput, StorageBrowserType } from './types';
+export {
+  CreateStorageBrowserInput,
+  StorageBrowserType,
+  DerivedActionViewType,
+  DerivedActionViews,
+} from './types';
+export { DefaultStorageBrowserDisplayText } from './displayText';