+++ packages/react-storage/src/components/StorageBrowser/providers/configuration/createConfigurationProvider.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { isComponent } from '@aws-amplify/ui-react-core/elements';
 
-import { ActionConfigsProvider } from '../../actions';
 import { CredentialsProvider } from './credentials';
 
 import { GetActionInputProvider } from './context';
@@ -19,7 +18,6 @@ export function createConfigurationProvider<T extends React.ComponentType<any>>(
 ): ConfigurationProviderComponent<T> {
   const {
     accountId,
-    actions,
     ChildComponent,
     displayName,
     region,
@@ -30,17 +28,15 @@ export function createConfigurationProvider<T extends React.ComponentType<any>>(
   const Child = isComponent(ChildComponent) ? ChildComponent : Passthrough;
 
   const Provider: ConfigurationProviderComponent<T> = (props) => (
-    <ActionConfigsProvider actions={actions}>
-      <CredentialsProvider {...rest}>
-        <GetActionInputProvider
-          accountId={accountId}
-          region={region}
-          customEndpoint={customEndpoint}
-        >
-          <Child {...props} />
-        </GetActionInputProvider>
-      </CredentialsProvider>
-    </ActionConfigsProvider>
+    <CredentialsProvider {...rest}>
+      <GetActionInputProvider
+        accountId={accountId}
+        region={region}
+        customEndpoint={customEndpoint}
+      >
+        <Child {...props} />
+      </GetActionInputProvider>
+    </CredentialsProvider>
   );
 
   Provider.displayName = displayName;