+++ packages/react-storage/src/components/StorageBrowser/views/LocationActionView/index.ts
@@ -1,26 +1,35 @@
-export { CopyView, CopyViewType, CopyViewState, useCopyView } from './CopyView';
+export {
+  CopyView,
+  CopyViewProps,
+  CopyViewType,
+  CopyViewState,
+  useCopyView,
+} from './CopyView';
 export {
   CreateFolderView,
+  CreateFolderViewProps,
   CreateFolderViewType,
   CreateFolderViewState,
   useCreateFolderView,
 } from './CreateFolderView';
 export {
   DeleteView,
+  DeleteViewProps,
   DeleteViewType,
   DeleteViewState,
   useDeleteView,
 } from './DeleteView';
+export { LocationActionView } from './LocationActionView';
 export {
   UploadView,
+  UploadViewProps,
   UploadViewType,
   UploadViewState,
   useUploadView,
 } from './UploadView';
 export { useActionView } from './useActionView';
-
 export {
-  LocationActionView,
+  ActionViewState,
+  LocationActionViewType,
   LocationActionViewProps,
-} from './LocationActionView';
-export { ActionViewState } from './types';
+} from './types';